차단 확인과 관리 목록을 최신 뮤트 흐름에 맞춘다
